Inventory/Inventory_Alloydb/Backend/requirements.txt
aiohttp==3.13.0
cloud-sql-python-connector==1.18.5
fastapi==0.121.2
google-cloud-alloydb-connector>=1.9.0
google-cloud-pubsub==2.31.1
google-cloud-storage>=2.18.2,<4.0.0
langchain==0.3.27
langchain-community==0.3.31
langchain-google-alloydb-pg==0.13.0
langchain-google-vertexai==2.0.9
#langchain_postgres==0.0.15
langchain-postgres>=0.0.15
langgraph==0.6.10
numpy==2.1.3
pandas==2.3.3
pg8000==1.31.5
prophet==1.2.1
psycopg[binary]==3.2.10
pydantic==2.9.2
python-dotenv==1.1.1
sqlalchemy==2.0.44
uvicorn[standard]==0.35.0
vertexai==1.71.1
plotly